Added rooms, floor maps, and room spawning
Rooms are a prefab that have 4 walls. Each wall can dynamically be
changed to having a door or not. If they have a door, the door can be
closed or opened.

Map represents the minimum amount of data to represent a room layout.
This includes a grid of MapEntries, each of which currently only has a
MapEntryType Type property that can be Empty, NormalRoom, SpawnRoom, BossRoom, or
ItemRoom.

RoomSpawner is responsible for instantiating rooms based on a map.

Floor will later be responsible for generating this map and using
RoomSpawner to spawn it in. Right now, it just spawns in a serialized
map.